# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Invillia Test",
        "description": "Invillia Test",
        "contact": {
            "email": "jonathanbeltrao@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/upload": {
            "post": {
                "tags": [
                    "Xml"
                ],
                "summary": "Upload XML",
                "description": "Upload given XML",
                "operationId": "uploadXml",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "customer_xml": {
                                        "type": "file"
                                    },
                                    "order_xml": {
                                        "type": "file"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "customer_xml": "customer.xml",
                                    "order_xml": "order.xml"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucess": {
                                            "description": "Status",
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "success": true
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucess": {
                                            "description": "Status",
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "success": false
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucess": {
                                            "description": "Status",
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "success": false
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}